[FIX] 입찰 기능 수정 및 @Authentication 수정 (#35)
* fix: 입찰 기능 수정 및 @Authentication 수정

* fix: local 에서 로그 저장할 때 로그 파일 최대 크기 수정

* fix: bidcode 메시지 수정

* fix: UserPrincipal 중복된 코드 수정

* fix: CreateAuctionRequest 에서 startprice를 long type으로 수정

* fix: CreateAuctionRequest 에서 startprice를 long type으로 수정

+++ omocha-client/src/main/java/org/auction/client/auction/interfaces/request/CreateAuctionRequest.java
@@ -13,7 +13,7 @@ public record CreateAuctionRequest(
 	@NotBlank String title,
 	String content,
 	// REFACTOR : 나중에 회의 후 VO를 도입할지 어떻게 처리할지 알아봐야함
-	@NotNull Integer startPrice,
+	@NotNull Long startPrice,
 	@NotNull Integer bidUnit,
 	@NotNull AuctionType auctionType,
 	@NotNull @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
